move shim building to `xtask`
